CogVM source as per VMMaker.oscog-eem.3102
CoInterpreter: fix a slip in printStringDataOf:on:.

SpurMemoryManager: move changeClassOf:to: up and use bytesPerOop to eliminate duplication of the whole method.

sqSCCSVersion.h: eliminate a warning

+++ platforms/Cross/vm/sqSCCSVersion.h
@@ -239,7 +239,7 @@ main(int _argc, char **_argv)
 	if (printit("VM_RELEASE"))
 		printf("%s\n", revisionAsString());
 	if (printit("VERSION_TAG"))
-		printf("%s-%s-%ibit\n", vm_version, revisionAsString(), 8 * sizeof(void *));
+		printf("%s-%s-%ibit\n", vm_version, revisionAsString(), (int)(8 * sizeof(void *)));
 	if (printit("VERSION_TAG_NO_BITS"))
 		printf("%s-%s\n", vm_version, revisionAsString());
 	if (printit("VERSION_NUMBER"))
